Fix service state indicator [WAL-2210]
Service state indicator is used in providers list component.
Indicator component accepts provider argument.
Provider is passed from provider list store by index.
However, when user reaches last page, usually it has only few elements.
Therefore provider is not defined for last elements as it refers to stale indexes.
As a result, undefined is passed as a provider and rendering fails.
In order to fix this issue we should not render Angular component for ReactJS
component if all of its bindings are undefined.